bug 1122658 - move get_ip to kuma.core.utils
return proper "60/m" rate string from IPBanJob.empty()

use IPBanJob().invalidate to refresh cache when deleting IPBan

move cache.clear to BannedIPTest.tearDown()
